pyproject.toml
[tool.poetry]
name = "pyside-demo"
version = "0.1.0"
description = "Offline-first GUI application with PostgreSQL synchronization"
repository = "https://github.com/jiriklic/pyside-demo"
authors = ["Jiri Klic <web@jiriklic.com>"]
license = "MIT"
readme = "README.md"
keywords = [
    "PySide6",
    "Offline-first",
    "GUI"
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pyside6 = "^6.7.2"
sqlalchemy = "^2.0.23"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
qtawesome = "^1.3.1"
folium = "^0.17.0"
pyqtgraph = "^0.13.7"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pymarkdownlnt = "^0.9.23"
pytest = "^8.3.3"
flake8-pyproject = "^1.2.3"
mypy = "^1.11.2"
types-psycopg2 = "^2.9.21.20240819"
types-requests = "^2.32.0.20240914"
pyside6-stubs = "^6.4.2.0"
sphinx = "7.4.7"
sphinx-rtd-theme = "^2.0.0"
nbsphinx = "^0.9.5"
myst-parser = "3.0.1"


[tool.black]
line-length = 79
force-exclude = "pyside_demo/resources"

[tool.isort]
profile = "black"
line_length = 79
skip = "pyside_demo/resources"

[tool.flake8]
exclude = [
    "pyside_demo/resources",
]

[tool.pymarkdown]
plugins.line-length.enabled = false
plugins.header-increment.enabled = false
plugins.no-duplicate-header.allow_different_nesting = true

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true
check_untyped_defs = true
exclude = [
    "pyside_demo/resources",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
